+++ index.js
@@ -8,20 +8,40 @@ var alexa = require('alexa-app');
 var bodyParser = require('body-parser');
 var alexaVerifierMiddleware = require('alexa-verifier-middleware');
 var Promise = require('bluebird');
+var defaults = require("lodash.defaults");
 
 var appServer = function(config) {
     var self = {};
     config = config || {};
-    var server_root = config.server_root || '';
 
-    if (config.verify === true && config.debug === true) {
+    var defaultOptions = {
+        log: true,
+        debug: true,
+        verify: false,
+        port: process.env.port || 8080,
+        httpEnabled: true,
+        httpsEnabled: false,
+        server_root: ''
+    };
+
+    config = defaults(config, defaultOptions);
+
+    if (config.verify && config.debug) {
         throw new Error("invalid configuration: the verify and debug options cannot be both enabled");
+    }
+
+    if (config.httpEnabled == false && config.httpsEnabled == false) {
+        throw new Error("invalid configuration: either http or https must be enabled");
+    }
+
+    if (config.httpEnabled && config.httpsEnabled && config.port == config.httpsPort) {
+        throw new Error("invalid configuration: http and https ports must be different");
     }
 
     self.apps = {};
 
     self.log = function(msg) {
-        if (config.log !== false) {
+        if (config.log) {
             console.log(msg);
         }
     };
@@ -110,7 +130,7 @@ var appServer = function(config) {
                         });
                 });
                 // Configure GET requests to run a debugger UI
-                if (false !== config.debug) {
+                if (config.debug) {
                     self.express.get(endpoint, function(req, res) {
                         if (typeof req.query['schema'] != "undefined") {
                             res.set('Content-Type', 'text/plain').send(app.schema());
@@ -195,7 +215,7 @@ var appServer = function(config) {
         }
 
         // Serve static content
-        var static_dir = path.join(server_root, config.public_html || 'public_html');
+        var static_dir = path.join(config.server_root, config.public_html || 'public_html');
         if (fs.existsSync(static_dir) && fs.statSync(static_dir).isDirectory()) {
             self.log("serving static content from: " + static_dir);
             self.express.use(express.static(static_dir));
@@ -204,7 +224,7 @@ var appServer = function(config) {
         }
 
         // Find any server-side processing modules and let them hook in
-        var server_dir = path.join(server_root, config.server_dir || 'server');
+        var server_dir = path.join(config.server_root, config.server_dir || 'server');
         if (fs.existsSync(server_dir) && fs.statSync(server_dir).isDirectory()) {
             self.log("loading server-side modules from: " + server_dir);
             self.load_server_modules(server_dir);
@@ -213,7 +233,7 @@ var appServer = function(config) {
         }
 
         // Find and load alexa-app modules
-        var app_dir = path.join(server_root, config.app_dir || 'apps');
+        var app_dir = path.join(config.server_root, config.app_dir || 'apps');
         if (fs.existsSync(app_dir) && fs.statSync(app_dir).isDirectory()) {
             self.log("loading apps from: " + app_dir);
             self.load_apps(app_dir, config.app_root || '/alexa/');
@@ -221,15 +241,13 @@ var appServer = function(config) {
             self.log("apps not loaded because directory [" + app_dir + "] does not exist");
         }
 
-        config.port = config.port || process.env.port || 8080;
-
-        if (config.https == true) {
+        if (config.httpsEnabled == true) {
             self.log("enabling https");
 
-            if (config.privateKey != undefined && config.certificate != undefined) { // Ensure that all of the needed properties are set
-                var privateKeyFile = server_root + '/sslcert/' + config.privateKey;
-                var certificateFile = server_root + '/sslcert/' + config.certificate;
-                var chainFile = (config.chain != undefined) ? server_root + '/sslcert/' + config.chain : undefined; //optional chain bundle
+            if (config.privateKey != undefined && config.certificate != undefined && config.httpsPort != undefined) { // Ensure that all of the needed properties are set
+                var privateKeyFile = config.server_root + '/sslcert/' + config.privateKey;
+                var certificateFile = config.server_root + '/sslcert/' + config.certificate;
+                var chainFile = (config.chain != undefined) ? config.server_root + '/sslcert/' + config.chain : undefined; //optional chain bundle
 
                 if (fs.existsSync(privateKeyFile) && fs.existsSync(certificateFile)) { // Make sure the key and cert exist.
                     var privateKey = fs.readFileSync(privateKeyFile, 'utf8');
@@ -267,11 +285,11 @@ var appServer = function(config) {
                             // TODO: add separate option to specify specific host address for HTTPS server to bind to???
                             // Issue #38: https://github.com/alexa-js/alexa-app-server/issues/38
                             if (typeof config.host === 'string') {
-                                self.instance = httpsServer.listen(config.port, config.host);
-                                self.log("listening on https://" + config.host + ":" + config.port);
+                                self.httpsInstance = httpsServer.listen(config.httpsPort, config.host);
+                                self.log("listening on https://" + config.host + ":" + config.httpsPort);
                             } else {
-                                self.instance = httpsServer.listen(config.port);
-                                self.log("listening on https port " + config.port);
+                                self.httpsInstance = httpsServer.listen(config.httpsPort);
+                                self.log("listening on https port " + config.httpsPort);
                             }
                         } catch (error) {
                             self.error("failed to listen via https: " + error);
@@ -283,9 +301,11 @@ var appServer = function(config) {
                     self.error("privateKey: '" + config.privateKey + "' or certificate: '" + config.certificate + "' do not exist in /sslcert, https will not be enabled");
                 }
             } else {
-                self.error("privatekey or certificate parameter is not set in config, https will not be enabled");
+                self.error("httpsPort, privateKey or certificate parameter is not set in config, https will not be enabled");
             }
-        } else {
+        }
+
+        if (config.httpEnabled) {
             if (typeof config.host === 'string') {
                 self.instance = self.express.listen(config.port, config.host);
                 self.log("listening on http://" + config.host + ":" + config.port);
@@ -300,13 +320,17 @@ var appServer = function(config) {
             config.post(self);
         }
 
-        return this.instance;
+        return this;
     };
 
     self.stop = function() {
         // close all server instances
         if (typeof self.instance !== "undefined") {
             self.instance.close();
+        }
+
+        if (typeof self.httpsInstance !== "undefined") {
+            self.httpsInstance.close();
         }
 
         // deactivate all hotswap listener