+++ test/test-examples-server-https-support-extended.js
@@ -20,15 +20,30 @@ describe("Alexa App Server with Examples & more HTTPS support", function() {
     testServer.stop();
   });
 
-  describe("No specific address given", function() {
-    it("should only have the HTTP server instance running", function() {
+  describe("no specific address given", function() {
+    it("has the HTTP server instance running on default port 8080", function() {
+      testServer = alexaAppServer.start({
+        server_root: 'examples'
+      });
+
+      return request(testServer.express)
+        .get('/alexa/hello_world')
+        .expect(200).then(function(response) {
+          expect(testServer.instance).to.exist;
+          return tcpPortUsed.check(8080).then(function(inUse) {
+            expect(inUse).to.equal(true);
+          });
+        });
+    });
+
+    it("has the HTTP server instance running on configured port 3000", function() {
       testServer = alexaAppServer.start({
         port: 3000,
         server_root: 'examples'
       });
 
       return request(testServer.express)
-        .get('/alexa/helloworld')
+        .get('/alexa/hello_world')
         .expect(200).then(function(response) {
           expect(testServer.instance).to.exist;
           return tcpPortUsed.check(3000).then(function(inUse) {
@@ -37,11 +52,11 @@ describe("Alexa App Server with Examples & more HTTPS support", function() {
         });
     });
 
-    it("should have an HTTPS server instances running", function() {
+    it("has the HTTPS server instances running", function() {
       testServer = alexaAppServer.start({
-        port: 3000,
+        httpsPort: 6000,
         server_root: 'examples',
-        https: true,
+        httpsEnabled: true,
         privateKey: 'private-key.pem',
         certificate: 'cert.cer',
         chain: 'cert.ca_bundle',
@@ -49,18 +64,19 @@ describe("Alexa App Server with Examples & more HTTPS support", function() {
       });
 
       return request(testServer.express)
-        .get('/alexa/helloworld')
+        .get('/alexa/hello_world')
         .expect(200).then(function(response) {
           expect(testServer.instance).to.exist;
-          return tcpPortUsed.check(3000).then(function(inUse) {
+          expect(testServer.httpsInstance).to.exist;
+          return tcpPortUsed.check(6000).then(function(inUse) {
             expect(inUse).to.equal(true);
           });
         });
     });
   });
 
-  describe("Specific address given (127.0.0.1)", function() {
-    it("should only have the HTTP server instance running", function() {
+  describe("on 127.0.0.1", function() {
+    it("has the HTTP server instance running", function() {
       testServer = alexaAppServer.start({
         port: 3000,
         host: '127.0.0.1',
@@ -68,21 +84,23 @@ describe("Alexa App Server with Examples & more HTTPS support", function() {
       });
 
       return request(testServer.express)
-        .get('/alexa/helloworld')
+        .get('/alexa/hello_world')
         .expect(200).then(function(response) {
           expect(testServer.instance).to.exist;
+          expect(testServer.httpsInstance).to.not.exist;
           return tcpPortUsed.check(3000).then(function(inUse) {
             expect(inUse).to.equal(true);
           });
         });
     });
 
-    it("should have an HTTPS server instances running", function() {
+    it("has both an HTTP and HTTPS server instances running", function() {
       testServer = alexaAppServer.start({
-        port: 6000,
+        port: 3000,
+        httpsPort: 6000,
         host: '127.0.0.1',
         server_root: 'examples',
-        https: true,
+        httpsEnabled: true,
         privateKey: 'private-key.pem',
         certificate: 'cert.cer',
         chain: 'cert.ca_bundle',
@@ -90,11 +108,43 @@ describe("Alexa App Server with Examples & more HTTPS support", function() {
       });
 
       return request(testServer.express)
-        .get('/alexa/helloworld')
+        .get('/alexa/hello_world')
         .expect(200).then(function(response) {
           expect(testServer.instance).to.exist;
+          expect(testServer.httpsInstance).to.exist;
           return tcpPortUsed.check(6000).then(function(inUse) {
             expect(inUse).to.equal(true);
+            return tcpPortUsed.check(3000).then(function(inUse) {
+              expect(inUse).to.equal(true);
+            });
+          });
+        });
+    });
+
+    it("with httpEnabled = false only starts an HTTPs instance", function() {
+      testServer = alexaAppServer.start({
+        httpEnabled: false,
+        port: 3000,
+        httpsPort: 6000,
+        host: '127.0.0.1',
+        server_root: 'examples',
+        httpsEnabled: true,
+        privateKey: 'private-key.pem',
+        certificate: 'cert.cer',
+        chain: 'cert.ca_bundle',
+        passphrase: "test123"
+      });
+
+      return request(testServer.express)
+        .get('/alexa/hello_world')
+        .expect(200).then(function(response) {
+          expect(testServer.instance).to.not.exist;
+          expect(testServer.httpsInstance).to.exist;
+          return tcpPortUsed.check(6000).then(function(inUse) {
+            expect(inUse).to.equal(true);
+            return tcpPortUsed.check(3000).then(function(inUse) {
+              expect(inUse).to.equal(false);
+            });
           });
         });
     });